+++ Changelog.md
@@ -7,6 +7,9 @@
 1. Add depth check and depth write properties to Material
     * [Pull request 160](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/160)
 
+1. Move scale functions from Visual to Node class.
+    * [Pull request 159](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/159)
+
 1. Add support for Render Passes and implement Gaussian Noise pass
     * [Pull request 154](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/154)
     * [Pull request 155](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/155)